+++ src/features/TodosViewForm/TodosViewForm.component.jsx
@@ -2,13 +2,16 @@ import './TodosViewForm.styles.css';
 
 function TodosViewForm() {
   return (
-    <form>
+    <form className="todos-view-form">
       <div className="form-controls">
         <label htmlFor="sort-by">Sort By</label>
         <select name="sort-by" id="sort-by">
           <option value="Title">Title</option>
           <option value="createdTime">Time Created</option>
         </select>
+
+        <span className="divider" aria-hidden="true" />
+
         <label htmlFor="direction">Direction</label>
         <select name="direction" id="direction">
           <option value="asc">Ascending</option>